update 3.0.6 fix bugs
- แก้ไขปัญหาเสียงกระซิบหาย
- แก้ไขปัญหาวัน-เวลาในหน้ารายการแชทส่วนตัวแสดงผลไม่ถูกต้อง
- แก้ไขปัญหาลบข้อความไม่หาย ต้อง Refresh (ในกรณีลบข้อความของตนเอง)
- แก้ไขให้ลบข้อความแล้วข้อความหายทันที แทนทีจะต้องรออัปเดตครั้งถัดไป (ในกรณีลบข้อความของตนเอง)
- แก้ไขตั้งค่าเสียงให้เปิดเสียงเป็นค่าเริ่มต้น
- แก้ไขข้อความในหน้าตั้งค่าปลั๊กอินใน AdminCP ให้สื่อความหมายให้เข้าใจง่ายขึ้น
- แก้ไขเวลาที่โพสต์ข้อความให้ติดข้างล่าง และไม่ตัดไปอีกบรรทัด
- แก้ไขดีไซน์อื่น ๆ เล็ก ๆ น้อย ๆ
- แก้ไขค่าเริ่มต้นของตั้งค่า "จำนวนข้อความที่จะแสดงเริ่มต้น" จาก 30 เป็น 50
- แก้ไขให้สามารถกำหนด "จำนวนข้อความในฐานข้อมูลสูงสุด" (เดิมใช้คำว่า "จำนวน log ที่จะเก็บไว้สูงสุด")
เป็น 0 เพื่อไม่จำกัดจำนวนข้อความในฐานข้อมูลได้ (โดยตั้งค่าเริ่มต้นเป็น 0 ให้หากติดตั้งแชทใหม่)
Developer Note: เนื่องจากแชทใหม่ที่แยกข้อความทั่วไปกับส่วนตัวจึงเห็นว่าไม่ควรลบข้อความเก่า ๆ ออกไป
และการเก็บข้อมูลแชทมาก ๆ แทบไม่มีผลต่อการใช้งานอย่างที่คิดกันในตอนแรก
แนะนำให้คนที่ใช้แชทเดิมแก้ไขให้เป็น 0 หลังจากอัปเกรดเป็นเวอร์ชั่นนี้

+++ th_chat/post.inc.php
@@ -155,7 +155,9 @@ $last = DB::insert_id();
 if ($needClear) {
     DB::query("DELETE FROM " . DB::table('newz_data') . " WHERE id<" . $last);
 } else {
-    DB::query("DELETE FROM " . DB::table('newz_data') . " WHERE id<" . ($last - $config['chat_log']));
+	if($config['chat_log']){
+		DB::query("DELETE FROM " . DB::table('newz_data') . " WHERE id<" . ($last - $config['chat_log']));
+	}
 }
 if ($_POST['touid']) {
     $room = intval($_POST['room']);
